requirements.txt
streamlit==1.45.1
streamlit-option-menu==0.4.0
requests==2.32.4
python-dotenv==1.1.0
databricks-sql-connector==4.0.3
pandas==2.3.0